requirements.txt
python-dotenv
openai
langchain
langchain-openai
pypdf2
unstructured
ipykernel
faiss-cpu
pypdf

streamlit